Rename scripts to corespond to the project name
In the past I called the project just nixdeploy but that was missleading
as it doesn't deploy Nix but rather NixOS and thus the name change was
in order. This finishes the previously only done repository rename to
also cover all scripts.

This also performs some refactoring of Nix code while I am at it.