system-script: fix usage with doas
doas requires command to be directly passed and doesn't support
additional environment variables. Thus let's first export and only after
that use sucmd.

doas can be configured to not pass environment variables but at worst
case scenario a closure diff will be printed again so that is not a huge
issue.