nixosdeploy.sh: handle the source access in better way
We want to make sure that in one invocation we use the same source no
matter what. Thus we first copy source to the store and use the store
path down the line. But url that was originally pointing to the store
path now points to the original source with some arguments added. We
need to use path instead to really get the store path.

This also removes shell configuration file that was never used nor
documented.